dbot-intent-service/requirements.txt
padatious==0.4.8
pyyaml==6.0
pytest==7.2.1
protobuf==4.22.0
